+++ package.json
@@ -16,6 +16,8 @@
     "react-dom": "^19.1.0",
     "react-responsive": "^10.0.1",
     "react-router-dom": "^7.6.3",
+    "react-swipeable": "^7.0.2",
+    "react-swipeable-list": "^1.10.0",
     "styled-components": "^6.1.19",
     "swiper": "^11.2.10",
     "ts-node": "^10.9.2"
